requirements.txt
Django==5.0
djangorestframework==3.14
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.12.0
debugpy==1.8.5
python-decouple==3.8
psycopg2==2.9.9
setuptools==74.1.2
requests==2.32.3
Pillow==10.4.0
django_jazzmin==3.0.1